+++ BankingSystem/User.h
@@ -2,15 +2,15 @@
 #include "MyString.h"
 
 namespace Constants{
-	const uint16_t MIN_AGE = 18;
-	const uint16_t MAX_AGE = 100;
+	const size_t MIN_AGE = 18;
+	const size_t MAX_AGE = 100;
 }
 
 class User {
 private:
 	MyString name="Unknown";
 	MyString egn="N/A";
-	uint16_t age=0;
+	size_t age=0;
 	MyString role="Unknown";
 	MyString password="N/A";
 
@@ -19,11 +19,11 @@ protected:
 
 public:
 	User() = default;
-	User(const MyString& name, const MyString& egn, uint16_t age, const MyString& role, const MyString& password);
+	User(const MyString& name, const MyString& egn, size_t age, const MyString& role, const MyString& password);
 
 	const MyString& getName() const;
 	const MyString& getEgn() const;
-	uint16_t getAge() const;
+	size_t getAge() const;
 	const MyString& getRole() const;
 	const MyString& getPassword() const;
